pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "intellichem2mqtt"
version = "1.0.0"
description = "Pentair IntelliChem to MQTT Bridge for Home Assistant"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Intellichem2MQTT Contributors"}
]
keywords = [
    "pentair",
    "intellichem",
    "mqtt",
    "home-assistant",
    "pool",
    "chemistry",
    "rs485",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: No Input/Output (Daemon)",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Home Automation",
]

dependencies = [
    "pyserial-asyncio>=0.6",
    "aiomqtt>=2.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
intellichem2mqtt = "intellichem2mqtt.__main__:main"

[project.urls]
Homepage = "https://github.com/yourusername/intellichem2mqtt"
Repository = "https://github.com/yourusername/intellichem2mqtt.git"
Issues = "https://github.com/yourusername/intellichem2mqtt/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["intellichem2mqtt*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py39"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.isort]
known-first-party = ["intellichem2mqtt"]
